// src/generated/models/Office365UsersModel.ts
export interface User {
  // A unique identifier for the user object in the directory.
  Id: string;
  // true if the account is enabled; otherwise, false. This property is required when a user is created.
  AccountEnabled?: boolean;
  // A list of business phone numbers for the user
  BusinessPhones?: string[];
  // The city in which the user is located.
  City?: string;
  // The name of the company in which the user works.
  CompanyName?: string;
  // The country/region in which the user is located; for example, "US" or "UK".
  Country?: string;
  // The name of the department in which the user works.
  Department?: string;
  // The name displayed in the address book for the user. This is usually the combination of the user's first name, middle initial and last name. This property is required when a user is created and it cannot be cleared during updates.
  DisplayName?: string;
  // The given name (first name) of the user.
  GivenName?: string;
  // The user’s job title.
  JobTitle?: string;
  // The SMTP/email address for the user, for example, "jeff@contoso.onmicrosoft.com". Read-Only.
  Mail?: string;
  // The mail alias for the user. This property must be specified when a user is created.
  MailNickname?: string;
  // The primary cellular telephone number for the user.
  mobilePhone?: string;
  // The office location in the user's place of business.
  OfficeLocation?: string;
  // The postal code for the user's postal address. The postal code is specific to the user's country/region. In the United States of America, this attribute contains the ZIP code.
  PostalCode?: string;
  // The user's surname (family name or last name).
  Surname?: string;
  // The primary cellular telephone number for the user.
  TelephoneNumber?: string;
  // The user principal name (UPN) of the user. The UPN is an Internet-style login name for the user based on the Internet standard RFC 822. By convention, this should map to the user’s email name. The general format is alias@domain, where domain must be present in the tenant’s collection of verified domains.
  UserPrincipalName?: string;
}
// ...
export interface GraphUser_V1 {
  // About Me
  aboutMe?: string;
  // true if the account is enabled; otherwise, false. This property is required when a user is created.
  accountEnabled?: boolean;
  // Birthday
  birthday?: string;
  // Business Phones
  businessPhones?: string[];
  // The city in which the user is located.
  city?: string;
  // The name of the company in which the user works.
  companyName?: string;
  // The country/region in which the user is located; for example, "US" or "UK".
  country?: string;
  // The name of the department in which the user works.
  department?: string;
  // Display Name
  displayName?: string;
  // Given Name
  givenName?: string;
  // Hire Date
  hireDate?: string;
  // Id
  id?: string;
  // Interests
  interests?: string[];
  // Job Title
  jobTitle?: string;
  // Mail
  mail?: string;
  // The mail alias for the user. This property must be specified when a user is created.
  mailNickname?: string;
  // Mobile Phone
  mobilePhone?: string;
  // My Site
  mySite?: string;
  // Office Location
  officeLocation?: string;
  // Past Projects
  pastProjects?: string[];
  // The postal code for the user's postal address. The postal code is specific to the user's country/region. In the United States of America, this attribute contains the ZIP code.
  postalCode?: string;
  // Preferred Language
  preferredLanguage?: string;
  // Preferred Name
  preferredName?: string;
  // Responsibilities
  responsibilities?: string[];
  // Schools
  schools?: string[];
  // Skills
  skills?: string[];
  // State
  state?: string;
  // Street Address
  streetAddress?: string;
  // Surname
  surname?: string;
  // User Principal Name
  userPrincipalName?: string;
  // User Type
  userType?: string;
}